requirements.txt
numpy
pandas
torch
accelerate
transformers
sentence-transformers
tokenizers
huggingface-hub
google-genai
ollama
neo4j
graphdatascience
networkx
chromadb
python-dotenv 
ipykernel 
rapidfuzz
pymupdf
pdf2image